+++ servo.c
@@ -35,8 +35,8 @@ void servo_destroy(struct servo *servo)
 }
 
 double servo_sample(struct servo *servo,
-		    double offset,
-		    double local_ts,
+		    int64_t offset,
+		    uint64_t local_ts,
 		    enum servo_state *state)
 {
 	return servo->sample(servo, offset, local_ts, state);
